Add files via upload
Adding a script that allows me to do this on my mac.  There was a write/read error. 

Connect the keyboard an push the little button to connect to mac.  Run script to connect the volume correctly for the keyboard.  
Then drag the files over, start with left then right.